requirements.txt
# Core ML & data
numpy>=1.25.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.12.0
joblib>=1.3.0

# Gradient boosting
catboost>=1.3.2

# Text features
sentence-transformers>=2.2.2
torch>=2.2.0
torchvision>=0.18.0
torchaudio>=2.2.0

# TF-IDF
scikit-learn>=1.3.0

# Image handling
Pillow>=10.1.0
tqdm>=4.66.0
requests>=2.32.0

# Optional for image embeddings (if using TensorFlow on M1)
tensorflow-macos>=2.15.0
tensorflow-metal>=0.9.0
